refactor: Header to move edit Icon feature outside
