+++ src/components/layouts/DefaultHeader.module.scss
@@ -7,29 +7,5 @@
   align-items: center;
   height: 60px;
   padding: 0 20px;
-}
-
-.userIcon {
-  width: 32px;
-  height: 32px;
-  background: #ddd;
-  vertical-align: bottom;
-}
-
-.spacing {
-  flex-grow: 1;
-}
-
-.nav {
-  padding: 8px;
-  border-bottom: 3px solid transparent;
-  transition: all 0.2s;
-}
-
-.nav:hover {
-  border-bottom: 3px solid #000000;
-}
-
-.padding {
-  height: 60px;
+  backdrop-filter: blur(1px);
 }
